manifest: Remove SchemaVersion field
SchemaVersion is never read - either to validate that it is the expected
version or to select different behavior. This removes the field to
simplify things - versioning will be handled by introducing new JSON
fields.

There are no backwards compatibility implications of this change - old
versions of Ollama will be fine with new manifests and new versions of
the software can handle old manifests.